Some changes to models
I created a Constant to include all the constants and enum classes, so
that they can be easier to find.

I plan to do some big changes to models by creating a base class called
model, which contains all the logic of file I/O, data loading and
conversion. If so, we don’t need to create a lot of model managers and
repeat redundant code in each model.